# terrasignal/frontend/openapi.json
{
  "components": {
    "schemas": {
      "AuditEventOut": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "actor_role": {
            "title": "Actor Role",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "event_id",
          "occurred_at",
          "actor",
          "actor_role",
          "event_type",
          "entity_type",
          "entity_id",
          "request_id",
          "payload"
        ],
        "title": "AuditEventOut",
        "type": "object"
      },
...
      "DriftMetricOut": {
        "properties": {
          "baseline_window": {
            "title": "Baseline Window",
            "type": "string"
          },
          "computed_at": {
            "format": "date-time",
            "title": "Computed At",
            "type": "string"
          },
          "current_window": {
            "title": "Current Window",
            "type": "string"
          },
          "feature_name": {
            "title": "Feature Name",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "psi": {
            "title": "Psi",
            "type": "number"
          },
          "status": {
            "enum": [
              "green",
              "amber",
              "red"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "model_name",
          "feature_name",
          "psi",
          "status",
          "computed_at",
          "baseline_window",
          "current_window"
        ],
        "title": "DriftMetricOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "KillSwitchRequest": {
        "properties": {
          "baseline_mode": {
            "title": "Baseline Mode",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "baseline_mode",
          "reason"
        ],
        "title": "KillSwitchRequest",
        "type": "object"
      },
      "KillSwitchState": {
        "properties": {
          "baseline_mode": {
            "title": "Baseline Mode",
            "type": "boolean"
          }
        },
        "required": [
          "baseline_mode"
        ],
        "title": "KillSwitchState",
        "type": "object"
      },
      "LineageOut": {
        "properties": {
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "as_of": {
            "format": "date",
            "title": "As Of",
            "type": "string"
          },
          "baseline_mode": {
            "title": "Baseline Mode",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dq_report_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dq Report Uri"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "eval_set_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Set Hash"
          },
          "features": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Features",
            "type": "object"
          },
          "feedback": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Feedback",
            "type": "array"
          },
          "git_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Sha"
          },
          "model_metrics": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Metrics"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "integer"
          },
          "output": {
            "additionalProperties": true,
            "title": "Output",
            "type": "object"
          },
          "prediction_id": {
            "format": "uuid",
            "title": "Prediction Id",
            "type": "string"
          },
          "training_snapshot_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Snapshot Uri"
          }
        },
        "required": [
          "prediction_id",
          "created_at",
          "model_name",
          "model_version",
          "entity_type",
          "entity_id",
          "as_of",
          "baseline_mode",
          "features",
          "output",
          "training_snapshot_uri",
          "dq_report_uri",
          "git_sha",
          "eval_set_hash",
          "model_metrics",
          "approved_by",
          "approved_at",
          "feedback"
        ],
        "title": "LineageOut",
        "type": "object"
      },
...
      "ModelVersionOut": {
        "properties": {
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "baseline_metrics": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "title": "Baseline Metrics",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dq_report_uri": {
            "title": "Dq Report Uri",
            "type": "string"
          },
          "eval_set_hash": {
            "title": "Eval Set Hash",
            "type": "string"
          },
          "git_sha": {
            "title": "Git Sha",
            "type": "string"
          },
          "metrics": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Metrics",
            "type": "object"
          },
          "model_card_path": {
            "title": "Model Card Path",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "training_snapshot_uri": {
            "title": "Training Snapshot Uri",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "model_name",
          "version",
          "status",
          "created_at",
          "metrics",
          "baseline_metrics",
          "eval_set_hash",
          "training_snapshot_uri",
          "dq_report_uri",
          "git_sha",
          "model_card_path",
          "approved_by",
          "approved_at"
        ],
        "title": "ModelVersionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "CRE rent forecasting & tenant default-risk platform.",
    "title": "TerraSignal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/governance/audit": {
      "get": {
        "operationId": "audit_trail_api_v1_governance_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "in": "query",
            "name": "actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AuditEventOut"
                  },
                  "title": "Response Audit Trail Api V1 Governance Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit Trail",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/governance/cards/{model_name}/{version}": {
      "get": {
        "operationId": "model_card_api_v1_governance_cards__model_name___version__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Model Card",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/governance/drift": {
      "get": {
        "operationId": "drift_api_v1_governance_drift_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DriftMetricOut"
                  },
                  "title": "Response Drift Api V1 Governance Drift Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Drift",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/governance/kill-switch": {
      "get": {
        "description": "Current baseline-mode flag. Analysts may read it (the UI banners on it);\nonly admins may flip it.",
        "operationId": "kill_switch_state_api_v1_governance_kill_switch_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillSwitchState"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Kill Switch State",
        "tags": [
          "governance"
        ]
      },
      "post": {
        "description": "Flip the API into (or out of) baseline mode without a redeploy. The flag\nlives in the DB so it survives a restart and takes effect on the next\nrequest; the flip is an audited action (\u00a78.6).",
        "operationId": "set_kill_switch_api_v1_governance_kill_switch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KillSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillSwitchState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Kill Switch",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/governance/lineage/{prediction_id}": {
      "get": {
        "operationId": "lineage_api_v1_governance_lineage__prediction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Prediction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineageOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Lineage",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/models/active": {
      "get": {
        "operationId": "models_active_api_v1_models_active_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ModelVersionOut"
                  },
                  "title": "Response Models Active Api V1 Models Active Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Models Active",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/models/{model_name}/versions/{version}/approve": {
      "post": {
        "description": "Human approval gate. Runs sync (registry helper) in the threadpool.",
        "operationId": "approve_api_v1_models__model_name__versions__version__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_name",
            "required": true,
            "schema": {
              "title": "Model Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Approve Api V1 Models  Model Name  Versions  Version  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve",
        "tags": [
          "governance"
        ]
      }
    },
...
  }
}